fix GOST
