+++ mapper/src/main/java/com/ibm/mapper/mapper/bc/BcBlockCipherMapper.java
@@ -41,8 +41,8 @@ import com.ibm.mapper.model.algorithms.SEED;
 import com.ibm.mapper.model.algorithms.Serpent;
 import com.ibm.mapper.model.algorithms.cast.CAST128;
 import com.ibm.mapper.model.algorithms.cast.CAST256;
-import com.ibm.mapper.model.algorithms.gost.Kuznyechik;
-import com.ibm.mapper.model.algorithms.gost.Magma;
+import com.ibm.mapper.model.algorithms.gost.GOST28147;
+import com.ibm.mapper.model.algorithms.gost.GOSTR34122015;
 import com.ibm.mapper.utils.DetectionLocation;
 import java.util.Optional;
 import javax.annotation.Nonnull;
@@ -76,8 +76,8 @@ public class BcBlockCipherMapper implements IMapper {
             case "DESedeEngine" -> Optional.of(new DESede(detectionLocation));
             case "DESEngine" -> Optional.of(new DES(detectionLocation));
             case "DSTU7624Engine" -> Optional.of(new Kalyna(detectionLocation));
-            case "GOST28147Engine" -> Optional.of(new Magma(detectionLocation));
-            case "GOST3412_2015Engine" -> Optional.of(new Kuznyechik(detectionLocation));
+            case "GOST28147Engine" -> Optional.of(new GOST28147(detectionLocation));
+            case "GOST3412_2015Engine" -> Optional.of(new GOSTR34122015(detectionLocation));
             case "IDEAEngine" -> Optional.of(new IDEA(detectionLocation));
             case "LEAEngine" -> Optional.of(new LEA(detectionLocation));
             case "NoekeonEngine" -> Optional.of(new NOEKEON(detectionLocation));
