+++ mapper/src/main/java/com/ibm/mapper/mapper/ssl/AuthenticationAlgorithmMapper.java
@@ -24,12 +24,12 @@ import com.ibm.mapper.model.Algorithm;
 import com.ibm.mapper.model.algorithms.DSS;
 import com.ibm.mapper.model.algorithms.ECCPWD;
 import com.ibm.mapper.model.algorithms.ECDSA;
-import com.ibm.mapper.model.algorithms.GOSTR341012;
 import com.ibm.mapper.model.algorithms.Kerberos;
 import com.ibm.mapper.model.algorithms.PSK;
 import com.ibm.mapper.model.algorithms.RSA;
 import com.ibm.mapper.model.algorithms.SHA;
 import com.ibm.mapper.model.algorithms.SHA2;
+import com.ibm.mapper.model.algorithms.gost.GOSTR341012;
 import com.ibm.mapper.utils.DetectionLocation;
 import java.util.Optional;
 import org.jetbrains.annotations.NotNull;
