+++ mapper/src/main/java/com/ibm/mapper/mapper/ssl/KeyExchangeAlgorithmMapper.java
@@ -25,11 +25,11 @@ import com.ibm.mapper.model.KeyAgreement;
 import com.ibm.mapper.model.algorithms.DH;
 import com.ibm.mapper.model.algorithms.ECCPWD;
 import com.ibm.mapper.model.algorithms.ECDH;
-import com.ibm.mapper.model.algorithms.GOSTR341112;
 import com.ibm.mapper.model.algorithms.Kerberos;
 import com.ibm.mapper.model.algorithms.PSK;
 import com.ibm.mapper.model.algorithms.RSA;
 import com.ibm.mapper.model.algorithms.SRP;
+import com.ibm.mapper.model.algorithms.gost.GOSTR341112;
 import com.ibm.mapper.utils.DetectionLocation;
 import java.util.Optional;
 import org.jetbrains.annotations.NotNull;
