--- mapper/src/main/java/com/ibm/mapper/model/algorithms/GOSTR34122015.java
@@ -1,48 +1,0 @@
-/*
- * SonarQube Cryptography Plugin
- * Copyright (C) 2024 IBM
- *
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to you under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.ibm.mapper.model.algorithms;
-
-import com.ibm.mapper.model.Algorithm;
-import com.ibm.mapper.model.AuthenticatedEncryption;
-import com.ibm.mapper.model.BlockCipher;
-import com.ibm.mapper.model.IPrimitive;
-import com.ibm.mapper.model.Mode;
-import com.ibm.mapper.utils.DetectionLocation;
-import javax.annotation.Nonnull;
-import org.jetbrains.annotations.NotNull;
-
-public final class GOSTR34122015 extends Algorithm implements BlockCipher, AuthenticatedEncryption {
-    private static final String NAME = "GOSTR34122015"; // Kuznyechik
-
-    public GOSTR34122015(@NotNull DetectionLocation detectionLocation) {
-        super(NAME, BlockCipher.class, detectionLocation);
-    }
-
-    public GOSTR34122015(@Nonnull Mode mode, @NotNull DetectionLocation detectionLocation) {
-        super(NAME, BlockCipher.class, detectionLocation);
-        this.put(mode);
-    }
-
-    public GOSTR34122015(
-            @Nonnull final Class<? extends IPrimitive> asKind,
-            @NotNull GOSTR34122015 gostr34122015) {
-        super(gostr34122015, asKind);
-    }
-}
